Update fortune.py
working with vector OS ver 1.6.0.3331, python 3.7.1
took off text input so you can just "ask" vector a question without typing it.
Of course he is not listening but still better than typing it.  Added some
sound and played around with the sleep timing.  the file format was not working for me
so moved them to main folder.  after doing this noticed the m.png was working in one
of the first few lines, so will probably go back and edit the png locations again.
but for now this is working for me.